Started adding Markdown preview window
Allows the user to preview the markdown for casualty narratives. Custom
buttons for adding pictures added, but currently do nothing. Should use
an alert() to get the filename of the picture, and then the caption.
Insterted then at the location the user had their cursor.

Using Meta controller for the AJAX preview, as numerous pages will need
this (updates, articles, etc).